+++ include/randombytes.h
@@ -3,6 +3,8 @@
 #ifndef randombytes_h
 #define randombytes_h
 
+#include <stdlib.h>
+
 /**
  * Randombytes initialization.
  * Initialization may be needed for some random number generators (e.g. CTR-DRBG).
@@ -23,6 +25,6 @@ void randombytes_init(unsigned char *entropy_input,
  * @param[in] xlen Number of random bytes to be generated
  * @return int 0 on success, -1 otherwise
  */
-int randombytes(unsigned char *x, unsigned long long xlen);
+int randombytes(unsigned char *x, size_t xlen);
 
 #endif /* randombytes_h */
