+++ src/common/randombytes_ctrdrbg.c
@@ -124,7 +124,7 @@ AES256_CTR_DRBG_Update(unsigned char *provided_data,
     memcpy(V, temp + 32, 16);
 }
 
-int randombytes(unsigned char *random_array, unsigned long long nbytes) {
+int randombytes(unsigned char *random_array, size_t nbytes) {
     int ret = randombytes_nist(random_array, nbytes);
 #ifdef ENABLE_CT_TESTING
     VALGRIND_MAKE_MEM_UNDEFINED(random_array, ret);
